// whirlpool/whirlpoolemea/cc-newsletter-popup-custom/react/typings/global.d.ts
export interface TimeSplit {
  hours: string;
  minutes: string;
  seconds: string;
}
export interface CheckboxProps {
  checkboxTitleEN?: string;
  checkboxTitleIT?: string;
  checkboxRequired?: boolean;
  checkboxNewsletter?: boolean;
  checkBoxErrorMessage_defaultLang: string;
  checkBoxErrorMessage_currentLang: string;
}